+++ app/models/spec_language.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 class SpecLanguage < ApplicationRecord
-  validates :name, uniqueness: true
+  validates :name, uniqueness: true, format: { with: /\A[a-zA-Z-]+\z/ }
 end